+++ index.php
@@ -1,53 +1,22 @@
 <?php
 
 require_once('templates/functions.php');
-
+require_once('templates/data.php');
 
 // показывать или нет выполненные задачи
-$show_complete_tasks = rand(0, 1);
-$projects = ['Входящие', 'Учеба', 'Работа', 'Домашние дела', 'Авто'];
-$tasks = [
-    [
-        'task' => 'Собеседование в IT компании',
-        'date' => '01.12.2019',
-        'category' => 'Работа',
-        'completed' => false
-    ],
-    [
-        'task' => 'Выполнить тестовое задание',
-        'date' => '25.12.2019',
-        'category' => 'Работа',
-        'completed' => false
-    ],
-    [
-        'task' => 'Сделать задание первого раздела',
-        'date' => '21.12.2019',
-        'category' => 'Учеба',
-        'completed' => true
-    ],
-    [
-        'task' => 'Встреча с другом',
-        'date' => '22.12.2019',
-        'category' => 'Входящие',
-        'completed' => false
-    ],
-    [
-        'task' => 'Купить корм для кота	',
-        'date' => 'null',
-        'category' => 'Домашние дела',
-        'completed' => false
-    ],
-    [
-        'task' => 'Заказать пиццу',
-        'date' => 'null',
-        'category' => 'Домашние дела',
-        'completed' => false
-    ]
-];
+
+    function countElements(array $elements, $value){
+        $intElement = 0;
+        foreach($elements as $element){
+            if($element['category'] == $value){
+                $intElement++;
+            }
+        };
+        return $intElement;
+    };
 
 
-
-    $page_content = include_template('main.php', ['projects' => $projects], ['tasks' => $tasks]);
+    $page_content = include_template('main.php', ['projects' => $projects, 'tasks' => $tasks]);
     $layout_content = include_template('layout.php', ['content', $page_content, 'title' => 'Дела в порядке']);
     print($layout_content);
